backend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
astroid==3.3.8
async-property==0.2.2
attrs==23.2.0
automat==22.10.0
billiard==4.2.1
black==24.3.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cli-exit-tools==1.2.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==41.0.7
dill==0.3.8
django==5.1.3
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-compat==1.0.15
django-cors-headers==4.4.0
django-filter==23.5
django-timezone-field==7.0
djangorestframework==3.15.2
docker==7.0.0
drf-spectacular==0.27.1
drf-standardized-errors==0.13.0
elastic-transport==8.15.1
elasticsearch==8.17.0
faker==24.2.0
flower==2.0.1
gunicorn==22.0.0
humanize==4.9.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
isort==6.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.4.2
lib-detect-testenv==2.0.8
markdown==3.5.2
mccabe==0.7.0
multiprocess==0.70.16
packaging==23.2
pathspec==0.12.1
pip==24.0
platformdirs==4.2.0
prometheus-client==0.20.0
prompt-toolkit==3.0.48
protobuf==5.27.3
psutil==5.9.8
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
pyasn1==0.6.0
pyasn1-modules==0.4.0
pycparser==2.22
pycurl==7.45.3
pygments==2.17.2
pylint==3.3.4
pyopenssl==24.1.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyyaml==6.0.1
redgreenunittest==0.1.1
redis==5.0.1
referencing==0.33.0
requests==2.31.0
responses==0.25.0
rpds-py==0.18.0
service-identity==24.1.0
setuptools==65.5.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.2
tblib==3.0.0
temporalio==1.6.0
tomlkit==0.13.2
tornado==6.4
twisted==24.3.0
txaio==23.1.1
types-protobuf==5.27.0.20240626
typing-extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.0
uv==0.4.2
uvloop==0.19.0
vine==5.1.0
watchdog==4.0.0
wcwidth==0.2.13
yapf==0.43.0
zope-interface==6.4.post2
gitdb==4.0.12
gitpython==3.1.44
smmap==5.0.2